submissions-durable-functions/requirements.txt
# Uncomment to enable Azure Monitor OpenTelemetry
# Ref: aka.ms/functions-azure-monitor-python 
# azure-monitor-opentelemetry 

azure-functions
azure-functions-durable
azure-cosmos
azure-identity
azure-storage-blob
azure-ai-documentintelligence
azure-servicebus
pydantic
python-dotenv
tenacity